some simple file restructuring
