+++ src/components/index.js
@@ -1,53 +1,53 @@
-import SidebarPlugin from "./SidebarPlugin";
-import Navbar from "./Navbar/Navbar";
-import NavbarToggleButton from "./Navbar/NavbarToggleButton";
-import Dropdown from "./Dropdown";
-import Card from "./Cards/Card";
-import StatsCard from "./Cards/StatsCard";
-import Table from "./Table";
 import AnimatedNumber from "./AnimatedNumber";
-import Progress from "./Progress";
 import Button from "./Button";
-import LoadingPanel from "./LoadingPanel";
-import Pagination from "./Pagination";
+import Card from "./Cards/Card";
+import Checkbox from "./Inputs/Checkbox";
 import Collapse from "./Collapse/Collapse";
 import CollapseItem from "./Collapse/CollapseItem";
+import Dropdown from "./Dropdown";
+import FormGroupInput from "./Inputs/FormGroupInput";
+import IconCheckbox from "./Inputs/IconCheckbox.vue";
+import LoadingPanel from "./LoadingPanel";
+import Pagination from "./Pagination";
+import Modal from "./Modal";
+import Navbar from "./Navbar/Navbar";
+import NavbarToggleButton from "./Navbar/NavbarToggleButton";
+import Notification from "./Notification";
+import Progress from "./Progress";
+import Radio from "./Inputs/Radio";
+import SidebarPlugin from "./SidebarPlugin";
+import Slider from "./Slider";
+import StatsCard from "./Cards/StatsCard";
+import Switch from "./Switch";
+import Table from "./Table";
 import TabPane from "./Tabs/TabPane";
 import Tabs from "./Tabs/Tabs";
-import Modal from "./Modal";
-import Notification from "./Notification";
-import Checkbox from "./Inputs/Checkbox";
-import Radio from "./Inputs/Radio";
-import FormGroupInput from "./Inputs/FormGroupInput";
 import VeeInput from "./Inputs/VeeInput";
-import Switch from "./Switch";
-import Slider from "./Slider";
-import IconCheckbox from "./Inputs/IconCheckbox.vue";
 
 export {
-  SidebarPlugin,
-  Navbar,
-  NavbarToggleButton,
-  Dropdown,
-  Card,
-  StatsCard,
-  Table,
   AnimatedNumber,
-  Progress,
   Button,
-  LoadingPanel,
-  Pagination,
+  Card,
+  Checkbox,
   Collapse,
   CollapseItem,
+  Dropdown,
+  FormGroupInput,
+  IconCheckbox,
+  LoadingPanel,
+  Pagination,
+  Modal,
+  Navbar,
+  NavbarToggleButton,
+  Notification,
+  Progress,
+  Radio,
+  SidebarPlugin,
+  Slider,
+  StatsCard,
+  Switch,
+  Table,
   TabPane,
   Tabs,
-  Modal,
-  Notification,
-  Checkbox,
-  Radio,
-  FormGroupInput,
   VeeInput,
-  Switch,
-  Slider,
-  IconCheckbox,
 };
